FIX_GR5-41_3

+++ projects/server/src/controllers/admin/category_admin.js
@@ -6,6 +6,10 @@ const product_category = db.Product_Category;
 module.exports = {
   addCategory: async (req, res) => {
     try {
+      if (req.role === 1 || req.role === 2) {
+        throw "Unauthorize Access";
+      }
+
       const { category } = req.body;
 
       await product_category.create({
@@ -20,6 +24,10 @@ module.exports = {
   },
   editCategory: async (req, res) => {
     try {
+      if (req.role === 1 || req.role === 2) {
+        throw "Unauthorize Access";
+      }
+
       await product_category.update(req.body, {
         where: {
           id: req.params.id,
@@ -34,6 +42,10 @@ module.exports = {
   },
   deleteCategory: async (req, res) => {
     try {
+      if (req.role === 1 || req.role === 2) {
+        throw "Unauthorize Access";
+      }
+
       await product_category.destroy({
         where: {
           id: req.params.id,
@@ -45,27 +57,12 @@ module.exports = {
       console.log(err);
     }
   },
-  // allCategory: async (req, res) => {
-  //   try {
-  //     const { sort, direction, pagination } = req.query;
-
-  //     const pages = Math.ceil((await product_category.count()) / 10);
-
-  //     const result = await product_category.findAll({
-  //       order: [[sort ? sort : "id", direction ? direction : "ASC"]],
-  //       limit: 10,
-  //       offset: pagination ? +pagination * 10 : 0,
-  //       raw: true,
-  //     });
-
-  //     res.status(200).send({ result, pages });
-  //   } catch (err) {
-  //     res.status(400).send(err);
-  //     console.log(err);
-  //   }
-  // },
   allCategory: async (req, res) => {
     try {
+      if (req.role === 1) {
+        throw "Unauthorize Access";
+      }
+
       const { search, sort, direction, pagination } = req.query;
 
       const pages = Math.ceil(
